% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\docType{class}
\name{cgdb-class}
\alias{cgdb-class}
\alias{cgdb}
\title{cgdb}
\description{
cgdb
}
\section{Slots}{

\describe{
\item{\code{db_root}}{}

\item{\code{cpgs}}{}

\item{\code{cells}}{}

\item{\code{plates}}{}
}}

