% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb-import.R
\name{cgdb_add_cell_from_df}
\alias{cgdb_add_cell_from_df}
\title{Add data from data frame to cgdb}
\usage{
cgdb_add_cell_from_df(df, db, cell_id, metadata = NULL, overwrite = TRUE,
  update_cells = TRUE)
}
\arguments{
\item{df}{data frame with the following fields: 
\itemize{
 \item{"cell_id" }{string with the cell id, needs to be of the format "{plate}.{num}", where num has to be numeric}
 \item{"chrom" }{chromosome, (character)}
 \item{"start" }{start position of the CpG (numeric)}
 \item{"end" }{end position of the CpG (start + 1, numeric)}
 \item{"cov" }{number of times the CpG was covered (numeric)}
 \item{"meth" }{number of times the CpG was methylated (numeric)}
 \item{"unmeth" }{number of times the CpG was unmethylated (cov - meth, numeric)}
}}

\item{db}{cgdb object}

\item{cell_id}{string with the cell id, needs to be of the format "{plate}.{num}", where num has to be numeric}

\item{metadata}{data frame with cell_id and additional fields with cell metadata (optional)}

\item{overwrite}{overwrite existing cells}

\item{update_cells}{update cells metadata for the db object}
}
\description{
Add data from data frame to cgdb
}
