% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb-import.R
\name{cgdb_add_cell_from_tidy_cpgs}
\alias{cgdb_add_cell_from_tidy_cpgs}
\title{Add data from tidy_cpgs data frame to cgdb}
\usage{
cgdb_add_cell_from_tidy_cpgs(
  cell_id,
  tidy_cpgs,
  db,
  metadata = NULL,
  single_cell = TRUE,
  overwrite = TRUE,
  update_cells = TRUE
)
}
\arguments{
\item{cell_id}{string with the cell id, needs to be of the format "{plate}.{num}", where num has to be numeric}

\item{tidy_cpgs}{tidy cpgs data frame}

\item{db}{cgdb object}

\item{metadata}{data frame with cell_id and additional fields with cell metadata (optional)}

\item{single_cell}{take only a single methylation call per CpG (chosen randomly)}

\item{overwrite}{overwrite existing cells}

\item{update_cells}{update cells metadata for the db object}
}
\description{
Add data from tidy_cpgs data frame to cgdb
}
