% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb-import.R
\name{cgdb_add_cell_from_tidy_cpgs_dir}
\alias{cgdb_add_cell_from_tidy_cpgs_dir}
\title{Add data from tidy_cpgs directory to cgdb}
\usage{
cgdb_add_cell_from_tidy_cpgs_dir(
  cell_id,
  tidy_cpgs_dir,
  db,
  tidy_cpgs_dir_non_unique = NULL,
  metadata = NULL,
  single_cell = TRUE,
  overwrite = TRUE,
  update_cells = TRUE
)
}
\arguments{
\item{cell_id}{string with the cell id, needs to be of the format "{plate}.{num}", where num has to be numeric}

\item{tidy_cpgs_dir}{path to the directory of the tidy_cpgs (unique)}

\item{db}{cgdb object}

\item{tidy_cpgs_dir_non_unique}{path to the directory of the non-unique tidy_cpgs for statistics computation (optionsl)}

\item{metadata}{data frame with cell_id and additional fields with cell metadata (optional)}

\item{single_cell}{take only a single methylation call per CpG (chosen randomly)}

\item{overwrite}{overwrite existing cells}

\item{update_cells}{update cells metadata for the db object}
}
\description{
Add data from tidy_cpgs directory to cgdb
}
