% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{cgdb_add_plate}
\alias{cgdb_add_plate}
\title{Add sc5mc data of a plate from smat object}
\usage{
cgdb_add_plate(cgdb, smat, plate_name = NULL, overwrite = TRUE)
}
\arguments{
\item{cgdb}{cgdb object}

\item{smat}{smat object}

\item{plate_name}{name of the plate}

\item{overwrite}{overwrite}
}
\description{
Add sc5mc data of a plate from smat object
}
