% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{cgdb_remove_plate}
\alias{cgdb_remove_plate}
\title{Remove plate from cgdb}
\usage{
cgdb_remove_plate(db, plate_name, force = FALSE)
}
\arguments{
\item{plate_name}{name of the plate}

\item{force}{force remove (no user prompt)}
}
\description{
Remove plate from cgdb
}
