% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{cgdb_save}
\alias{cgdb_save}
\title{Write cgdb to disk}
\usage{
cgdb_save(db, path = NULL, force = FALSE)
}
\arguments{
\item{db}{cgdb object}

\item{path}{path of the cgdb view}

\item{force}{force overwrite}
}
\description{
Write cgdb to disk
}
