% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{extract.cgdb}
\alias{extract.cgdb}
\title{Extract data from intervals and cells}
\usage{
\method{extract}{cgdb}(.Object, cells = NULL, cpgs = NULL)
}
\arguments{
\item{cells}{cells to extract}

\item{cpgs}{cpgs to extract}
}
\description{
Extract data from intervals and cells
}
