% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{extract_sparse}
\alias{extract_sparse}
\title{Extract data from intervals and cells in a sparse format}
\usage{
extract_sparse(.Object, cells = NULL, cpgs = NULL)
}
\arguments{
\item{cells}{cells to extract}

\item{cpgs}{cpgs to extract}
}
\description{
Extract data from intervals and cells in a sparse format
}
