% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{filter_by_cov}
\alias{filter_by_cov}
\title{Filter cgdb object by cpgs / cells coverage}
\usage{
filter_by_cov(db, min_cpgs = 0, max_cpgs = Inf, min_cells = 0, max_cells = Inf)
}
\arguments{
\item{db}{smat object}

\item{min_cpgs}{minimal number of CpGs}

\item{max_cpgs}{maximal number of CpGs}

\item{min_cells}{minimal number of cells}

\item{max_cells}{maximal number of cells}
}
\description{
Filter cgdb object by cpgs / cells coverage
}
