% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{intervals_cell_cov}
\alias{intervals_cell_cov}
\title{Calculate number of cells per interval}
\usage{
intervals_cell_cov(db, intervals, min_cov = 1)
}
\arguments{
\item{db}{cgdb object}

\item{intervals}{intervals set to summarise the genome with. Numeric values would calculate
the number of cells for equally sized bins.}

\item{min_cov}{minimal coverage per cell}
}
\value{
number of cells that were covered (coverage >= \code{min_cov}) for the intervals set.
}
\description{
Calculate number of cells per interval
}
