% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-cycle.R
\name{phase_cells}
\alias{phase_cells}
\title{Phase cells (to G and S) using ratio of coverage in late and early replicating regions}
\usage{
phase_cells(db, early_tor, late_tor, ratio_thresh, tor_track = NULL)
}
\arguments{
\item{db}{cgdb object}

\item{early_tor}{early_tor}

\item{late_tor}{late_tor}

\item{ratio_thresh}{ratio_thresh}

\item{tor_track}{tor_track}
}
\description{
Phase cells (to G and S) using ratio of coverage in late and early replicating regions
}
\examples{
\dontrun{
		 cell_phases <- phase_cells(db, late_tor=c(0, 50), early_tor=c(66, 83), ratio_thresh=1.42, tor_track='encode.repliseq.wgEncodeUwRepliSeqImr90WaveSignalRep1')
}

  
}
