% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpdb.R
\docType{class}
\name{plpdb-class}
\alias{plpdb-class}
\alias{plpdb}
\title{plpdb}
\description{
plpdb
}
\section{Slots}{

\describe{
\item{\code{db_root}}{}

\item{\code{cpgs}}{}

\item{\code{cells}}{}

\item{\code{CPG_NUM}}{}
}}

