% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpdb.R
\name{plpdb_add_plate}
\alias{plpdb_add_plate}
\title{Add plate to plpdb frpm smat reads}
\usage{
plpdb_add_plate(
  db,
  prefix,
  plate_name = NULL,
  overwrite = TRUE,
  update_cells = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{prefix}{path prefix of the location of the smat files}
}
\description{
Add plate to plpdb frpm smat reads
}
