% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpdb.R
\name{plpdb_init}
\alias{plpdb_init}
\title{Initialize pileup database}
\usage{
plpdb_init(db_root, binsize, overwrite = FALSE)
}
\arguments{
\item{db_root}{root directory of the cgdb database}

\item{binsize}{size of genomic bins}

\item{overwrite}{overwrite existing db}

\item{cgdb}{cgdb object to initialize from}
}
\description{
Initialize pileup database
}
