% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpdb.R
\name{plpdb_load}
\alias{plpdb_load}
\title{Load plpdb database from disk}
\usage{
plpdb_load(db_root)
}
\arguments{
\item{db_root}{root directory of the cgdb database}
}
\description{
Load plpdb database from disk
}
