% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{sc5mc.calc_pdiff_cor}
\alias{sc5mc.calc_pdiff_cor}
\title{Calculate pairwise correlation of cells pdiff}
\usage{
sc5mc.calc_pdiff_cor(cell_comp, min_cells, use = "pairwise.complete.obs",
  method = "spearman", cells = NULL)
}
\arguments{
\item{cell_comp}{Output of \code{sc5mc.calc_pdiff}: data frame with pairs of cells (cell1, cell2) and 'pdiff' column with the fraction of different CpGs.}

\item{min_cells}{minimal number of cells with pdiff statistic in both cells}

\item{use}{`use` parameter of `cor` function}

\item{method}{`method` parameter of `cor` function}

\item{cells}{vector with cell names of cells to operate on. If NULL calculation would be 
done on all cells.}
}
\description{
Calculate pairwise correlation of cells pdiff
}
