% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{sc5mc.calc_pdiff_rows}
\alias{sc5mc.calc_pdiff_rows}
\title{calculate pairwise statistics of CpGs}
\usage{
sc5mc.calc_pdiff_rows(smat, min_cells = 5, intervs = NULL, cols = NULL)
}
\arguments{
\item{smat}{smat object}

\item{min_cells}{minimal number of cells}

\item{intervs}{}

\item{cols}{}
}
\description{
calculate pairwise statistics of CpGs
}
