% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.cells_per_umi}
\alias{sc5mc.cells_per_umi}
\title{Calculates cells per UMI}
\usage{
sc5mc.cells_per_umi(smat)
}
\arguments{
\item{smat}{smat object}
}
\value{
tibble with number of cells ('n_cells') and number of molecules ('n')
}
\description{
Calculates cells per UMI
}
