% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{sc5mc.cluster_cells}
\alias{sc5mc.cluster_cells}
\title{Cluster cells}
\usage{
sc5mc.cluster_cells(smat, min_cells = 10, min_cgs = 100,
  use = "pairwise.complete.obs", method = "spearman",
  clustering_distance = "euclidean", clustering_method = "ward.D2",
  intervs = NULL, cols = NULL, samp_data = NULL, plot = FALSE, ...)
}
\arguments{
\item{smat}{smat object}

\item{min_cells}{minimal number of cells with pdiff statistic in both cells}

\item{min_cgs}{minimal number of cpgs covered in both cells}

\item{use}{`use` parameter of `cor` function}

\item{method}{`method` parameter of `cor` function}

\item{clustering_distance}{distance measure used for clustering the correlations}

\item{clustering_method}{clustering method used for clustering the correlations. Accepts the same values as ‘hclust’.}

\item{intervs}{intervals set to filter smat by before calculating. if NULL
all CpGs would be used.}

\item{cols}{vector of names of cells to use. If NULL all cells would be used.}

\item{samp_data}{data frame with additional annotation of the cell.
must have 'track' field with the cell name.}

\item{plot}{plot the clustered correlation matrix}

\item{...}{additional parameters for sc5mc.plot_cor_mat}
}
\value{
list with the following components: 
'cell_comp' - pairwise comparison of cell CpGs.
'cell cor' - pairwise cell correlations.
'mat' - correlation matrix.
'hclust_rows' - hclust object used to cluster the rows.
'hclust_cols' - hclust object used to cluster the columns.
If plot is TRUE the list is returned invisibly
}
\description{
Cluster cells
}
