% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{sc5mc.generate_metacells}
\alias{sc5mc.generate_metacells}
\title{Genarate metacells from TAD methylation}
\usage{
sc5mc.generate_metacells(
  db,
  tad_intervs,
  knn = 80,
  k_expand = 10,
  min_cluster_size = 20,
  n_resamp = 1000,
  seed = NULL,
  K = 50,
  min_mc_size = 30,
  min_tad_cov = 10,
  min_tad_cells = 50,
  min_tads = 1000,
  tads_meth = NULL,
  rm_xy = FALSE,
  normalize_tads = TRUE,
  downsample = FALSE,
  min_cov = NULL,
  ds_quantile = 0.1,
  norm_knn_mat = NULL,
  norm_knn_k = 20
)
}
\arguments{
\item{db}{cgdb object}

\item{tad_intervs}{intervals set of TADs}

\item{knn}{knn}

\item{k_expand}{k_expand}

\item{min_cluster_size}{min_cluster_size}

\item{n_resamp}{n_resamp}

\item{seed}{seed}

\item{K}{K}

\item{min_mc_size}{min_mc_size}

\item{min_tad_cov}{min_tad_cov}

\item{min_tad_cells}{min_tad_cells}

\item{min_tads}{min_tads}

\item{tads_meth}{previously computed methylation of TADS (using get_tads_meth)}

\item{rm_xy}{remove TADs from X and Y chromosome}

\item{normalize_tads}{normalize TAD methylation by dividing by the sum of each row}

\item{downsample}{downsample cells to have the same coverage}

\item{min_cov}{minimial coverage per cell (for downsampling). If NULL the value would be determined by the coverage of the cell in \code{ds_quantile} quantile}

\item{ds_quantile}{see \code{min_cov}. Quantile of the cell that determines the minimal coverage for downsampling.}

\item{norm_knn_mat}{matrix to use for normalizing orthogonal effects like cell cycle. Matrix should have a row per cell and a column per variable to normalize. For example, the matrix can have one column of coverage in early/coverage in late and another column of difference in methylation in early vs late regions.}

\item{norm_knn_k}{k parameter for normalization}
}
\value{
list with the following fields: mc_map, tads_clust, tads_meth, tads_meth_f, tads_mc
}
\description{
Genarate metacells from TAD methylation
}
