% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.index_plots}
\alias{sc5mc.index_plots}
\title{plot indexes statistics}
\usage{
sc5mc.index_plots(smat, ofn = NULL, width = 7 * 2.5, height = 5 * 2.3,
  cells_per_umi = FALSE, return_plotlist = FALSE, min_reads = 1000)
}
\arguments{
\item{smat}{smat object}

\item{ofn}{figure filename (ggsave parameter)}

\item{width}{figure width (ggsave parameter)}

\item{height}{figure height (ggsave parameter)}

\item{cells_per_umi}{plot cells per umi}

\item{return_plotlist}{return a list of plots}
}
\value{
plots of total reads, %mapping and %unique
}
\description{
plot indexes statistics
}
