% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.plot_cell_marginals}
\alias{sc5mc.plot_cell_marginals}
\title{plot cell marginals}
\usage{
sc5mc.plot_cell_marginals(smat, type = "percent")
}
\arguments{
\item{smat}{smat object}

\item{type}{'percent' or 'vals'}
}
\description{
plot cell marginals
}
