% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.plot_cell_pairs_coverage}
\alias{sc5mc.plot_cell_pairs_coverage}
\title{plot cell pairs coverage}
\usage{
sc5mc.plot_cell_pairs_coverage(smat)
}
\arguments{
\item{smat}{smat object}
}
\description{
plot cell pairs coverage
}
