% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.plot_cells_per_umi}
\alias{sc5mc.plot_cells_per_umi}
\title{Plots cells per UMI}
\usage{
sc5mc.plot_cells_per_umi(smat)
}
\arguments{
\item{smat}{smat object}
}
\value{
plot of number of cells versus percent of molecules (only for molecules that appeared in more than 1 cell)
}
\description{
Plots cells per UMI
}
