% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.plot_cg_pairs_coverage}
\alias{sc5mc.plot_cg_pairs_coverage}
\title{plot cg pairs coverage}
\usage{
sc5mc.plot_cg_pairs_coverage(smat, min_cells = 30)
}
\arguments{
\item{smat}{smat object}
}
\description{
plot cg pairs coverage
}
