% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{sc5mc.plot_cpg_marginals}
\alias{sc5mc.plot_cpg_marginals}
\title{plot cpg marginals}
\usage{
sc5mc.plot_cpg_marginals(smat, type = "percent")
}
\arguments{
\item{smat}{smat object}

\item{type}{'percent' or 'vals'}
}
\description{
plot cpg marginals
}
