% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{sc5mcs.calc_pdiff_cor_rows}
\alias{sc5mcs.calc_pdiff_cor_rows}
\title{calculate pairwise correlation of CpGs pdiff}
\usage{
sc5mcs.calc_pdiff_cor_rows(
  row_comp,
  min_intervals,
  pairwise.complete.obs = TRUE,
  spearman = TRUE,
  intervals = NULL
)
}
\arguments{
\item{row_comp}{}

\item{min_intervals}{min_intervals}

\item{pairwise.complete.obs}{pairwise.complete.obs}

\item{spearman}{spearman}

\item{intervals}{intervals}
}
\description{
calculate pairwise correlation of CpGs pdiff
}
