% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{smat.calc_pdiff}
\alias{smat.calc_pdiff}
\title{calculate pairwise statistics of cells}
\usage{
smat.calc_pdiff(smat, min_cgs = 100, intervs = NULL, cols = NULL,
  samp_data = NULL)
}
\arguments{
\item{smat}{smat object}

\item{min_cgs}{minimal number of cpgs covered in both cells}

\item{intervs}{intervals set to filter smat by before calculating. if NULL
all CpGs would be used.}

\item{cols}{vector of names of cells to use. If NULL all cells would be used.}

\item{samp_data}{data frame with additional annotation of the cell.
must have 'track' field with the cell name.}
}
\value{
smat object with cell_comp field
}
\description{
calculate pairwise statistics of cells
}
