% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{smat.cluster_cells}
\alias{smat.cluster_cells}
\title{Cluster cells}
\usage{
smat.cluster_cells(smat, min_cells = 10, min_cgs = 100,
  pairwise.complete.obs = TRUE, spearman = TRUE,
  clustering_method = "ward.D2", intervs = NULL, cols = NULL,
  samp_data = NULL, plot = FALSE, ...)
}
\arguments{
\item{smat}{smat object}

\item{min_cells}{minimal number of cells with pdiff statistic in both cells}

\item{min_cgs}{minimal number of cpgs covered in both cells}

\item{pairwise.complete.obs}{if TRUE: similar to use = 'pairwise.complete.obs' in \code{cor}. if FALSE: similar to use = 'everything'}

\item{spearman}{if 'TRUE' Spearman correlation is computed, otherwise Pearson}

\item{clustering_method}{clustering method used for clustering the correlations. Accepts the same values as ‘hclust’.}

\item{intervs}{intervals set to filter smat by before calculating. if NULL
all CpGs would be used.}

\item{cols}{vector of names of cells to use. If NULL all cells would be used.}

\item{samp_data}{data frame with additional annotation of the cell.
must have 'track' field with the cell name.}

\item{plot}{plot the clustered correlation matrix}

\item{...}{additional parameters for sc5mc.plot_cor_mat}
}
\value{
smat object with the following fields: 
\describe{
  \item{cell_comp:}{pairwise comparison of cell CpGs.}
  \item{cell_cor:}{pairwise cell correlations.}
  \item{cell_cor_mat:}{cells correlation matrix.}
  \item{cell_cor_hclust_rows:}{hclust object used to cluster the rows.}
  \item{cell_cor_hclust_cols:}{hclust object used to cluster the columns.}
}
If plot is TRUE the smat object is returned invisibly
}
\description{
Cluster cells
}
