% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.cpg_avg_marginals}
\alias{smat.cpg_avg_marginals}
\title{calculate marginal average methylation of CpGs}
\usage{
smat.cpg_avg_marginals(smat, cols = NULL)
}
\arguments{
\item{smat}{smat object}

\item{cols}{column names of columns to use. If NULL all columns (cells) would be returned}
}
\value{
intervals set (chrom, start, end) with the cpgs, and `cov`, `meth`, `unmeth` and `avg`
additional fields
}
\description{
calculate marginal average methylation of CpGs
}
