% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.cpg_pairs_marginals}
\alias{smat.cpg_pairs_marginals}
\title{Calculate the joint coverage of all pairs of CpGs}
\usage{
smat.cpg_pairs_marginals(smat, min_cells = 30, cols = NULL)
}
\arguments{
\item{smat}{smat object}

\item{min_cells}{minimal number of cells per CpG}

\item{cols}{column names of columns to use. If NULL all columns (cells) would be returned}
}
\value{
tibble with CpG coverage ('ncells') and the number of CpG pairs with that number 
jointly covered cells ('npairs')
}
\description{
Calculate the joint coverage of all pairs of CpGs
}
