% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.filter}
\alias{smat.filter}
\title{Filter smat object}
\usage{
smat.filter(smat, intervs = NULL, cols = NULL, ids = NULL)
}
\arguments{
\item{smat}{smat object}

\item{intervs}{intervals set (df with chrom,start,end fields) to filter smat by.
only CpGs covered by the intervals would be returned. if NULL all CpGs would be returned.}

\item{cols}{column names of columns to keep. If NULL all columns (cells) would be returned}

\item{ids}{vector ids of CpGs to filter by.}
}
\value{
smat object with CpGs covered by intervs, and cells that in cols / ids.
}
\description{
Filter smat object
}
