% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat_import.R
\name{smat.from_bams}
\alias{smat.from_bams}
\title{Create smat object from mapped reads (bam files)}
\usage{
smat.from_bams(metadata, groot, prefix = NULL, workdir = tempdir(),
  use_sge = TRUE, name = "", description = "",
  keep_tidy_cpgs = TRUE, load_existing = FALSE, filter_dups = FALSE,
  io_saturation = FALSE, threads = 10, cell_metadata = NULL,
  keep_tidy_calls = TRUE, single_cell = TRUE, ...)
}
\arguments{
\item{metadata}{data frame with 'lib' column with the library name (ususally cell id)
and 'bam' column with bam files. Multiple bams of the same cell should appear in separate rows.}

\item{groot}{misha db path}

\item{prefix}{path prefix in which to save the smat files. if NULL the object
would not be written to disk}

\item{workdir}{temporary directory to use for bam parsing}

\item{use_sge}{use sun grid engine cluster}

\item{name}{name of the object}

\item{description}{description of the object}

\item{keep_tidy_cpgs}{keep tidy cpgs attached to the smat object}

\item{load_existing}{load existing tidy cpgs}

\item{filter_dups}{relevenat only when load_existing is TRUE: filter duplicates}

\item{single_cell}{take only a single methylation call per CpG}

\item{...}{additional parameters to \code{gpatterns::gpatterns.import_from_bam}}
}
\value{
smat object (inivisibly if prefix is not NULL)
}
\description{
Create smat object from mapped reads (bam files)
}
