% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.from_conf}
\alias{smat.from_conf}
\title{Load smat object from disk}
\usage{
smat.from_conf(conf)
}
\arguments{
\item{conf}{list with 'sparse_matrices' field that holds a list with:
cov - path of coverage sparse matrix
meth - path of meth sparse matrix
unmeth - path of unmeth sparse matrix
intervs - path of intervs file
colnames - path of colnames file}
}
\value{
smat object
}
\description{
Load smat object from disk
}
