% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat_import.R
\name{smat.from_df}
\alias{smat.from_df}
\title{Create smat object from data frame}
\usage{
smat.from_df(df, name = "", description = "")
}
\arguments{
\item{df}{intervals ('chrom', 'start', 'end' fields) with additional fields:
meth (methylated calls), unmeth (unmethylated calls) and cov (total coverage)}

\item{name}{name of the object}

\item{description}{description of the object}
}
\value{
smat object
}
\description{
Create smat object from data frame
}
