% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat_import.R
\name{smat.from_tracks}
\alias{smat.from_tracks}
\title{Create smat object from misha tracks}
\usage{
smat.from_tracks(tracks, libs, prefix = NULL, use_sge = TRUE, ...)
}
\arguments{
\item{tracks}{names of tracks}

\item{libs}{names of new object smat columns}

\item{prefix}{path prefix in which to save the smat files. if NULL the object
would not be written to disk}

\item{use_sge}{use sun grid engine cluster}

\item{...}{additional parameters to gcluster.run2}
}
\value{
smat object
}
\description{
Create smat object from misha tracks
}
