% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.get_tidy_cpgs}
\alias{smat.get_tidy_cpgs}
\title{Get tidy cpgs of smat object}
\usage{
smat.get_tidy_cpgs(smat, unique = TRUE)
}
\arguments{
\item{smat}{smat object}

\item{unique}{if TRUE - get unique tidy_cpgs}
}
\value{
smat with additional tidy_cpgs field, that contain tidy_cpgs with additional 'cell_id' column with the cell id
}
\description{
Get tidy cpgs of smat object
}
