% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.join}
\alias{smat.join}
\title{join two smat objects}
\usage{
smat.join(smat_r, smat_l, type = "full", prefix_l = "l_", prefix_r = "r_")
}
\arguments{
\item{smat_r}{first smat object}

\item{smat_l}{seconf smat object}

\item{type}{join type. can be 'inner', 'full', 'left', 'right'. see xx_join by dplyr package.}

\item{prefix_l}{prefix for cell names from first smat in case of conflicts}

\item{prefix_r}{prefix for cell names from second smat in case of conflicts}
}
\value{
smat object with intervals joined by \code{type}_join
}
\description{
join two smat objects
}
