% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.load}
\alias{smat.load}
\title{Load smat object from disk}
\usage{
smat.load(prefix)
}
\arguments{
\item{prefix}{path prefix of the location of the smat files}
}
\value{
smat object
}
\description{
Load smat object from disk
}
