% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.multi_join}
\alias{smat.multi_join}
\title{join multiple smat objects}
\usage{
smat.multi_join(..., type = "full", prefix_l = "l_", prefix_r = "r_")
}
\arguments{
\item{...}{smat objects to join}

\item{type}{join type. can be 'inner', 'full', 'left', 'right', 'semi' and 'anti'. uses xx_join by dplyr package.}

\item{prefix_l}{prefix for cell names from first smat in case of conflicts}

\item{prefix_r}{prefix for cell names from second smat in case of conflicts}
}
\value{
joined smat objects
}
\description{
join multiple smat objects
}
