% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.multi_join_from_files}
\alias{smat.multi_join_from_files}
\title{join multiple smat objects from path}
\usage{
smat.multi_join_from_files(paths, type = "full")
}
\arguments{
\item{paths}{path prefixes of the location of the smats files}

\item{type}{join type. can be 'inner', 'full', 'left', 'right'. see xx_join by dplyr package.}
}
\value{
joined smat object
}
\description{
join multiple smat objects from path
}
