% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{smat.plot_cor_mat}
\alias{smat.plot_cor_mat}
\title{plot cell cell correlation matrix}
\usage{
smat.plot_cor_mat(smat, min_vals_row = 100, min_vals_col = 100,
  row_ord = NULL, col_ord = NULL, show_colnames = FALSE,
  show_rownames = FALSE, breaks = NULL, color_pal = NULL, ...)
}
\arguments{
\item{smat}{smat object}

\item{...}{}
}
\description{
plot cell cell correlation matrix
}
