% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.save}
\alias{smat.save}
\title{Save smat object to disk}
\usage{
smat.save(smat, prefix)
}
\arguments{
\item{smat}{smat object}

\item{prefix}{path prefix of the location of the smat files}
}
\value{
inivisibly returns the smat object
}
\description{
Save smat object to disk
}
