% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.select}
\alias{smat.select}
\title{Select cells from smat object}
\usage{
smat.select(smat, cols = NULL)
}
\arguments{
\item{smat}{smat object}

\item{cols}{column names of columns to keep. If NULL all columns (cells) would be returned}
}
\description{
Select cells from smat object
}
