% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.summarise}
\alias{smat.summarise}
\title{Summarise groups of CpGs per cell}
\usage{
smat.summarise(smat, groups, group_name = "group", parallel = TRUE)
}
\arguments{
\item{smat}{sc5mc smat object}

\item{groups}{character / factor vector with a group per CpG}

\item{group_name}{column name of the group}

\item{parallel}{compute parallely per group (using doMC package)}
}
\value{
data frame with the following fields:
'group_name', 'cell_id', 'ncpgs', 'meth', 'unmeth', 'avg'
}
\description{
Summarise groups of CpGs per cell
}
