% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.summarise_by_intervals}
\alias{smat.summarise_by_intervals}
\title{Group CpGs by intervals, and summarise per cell}
\usage{
smat.summarise_by_intervals(smat, intervals, return_smat = FALSE)
}
\arguments{
\item{smat}{smat object}

\item{intervals}{intervals set. CpGs covered by each interval would be summarised}

\item{return_smat}{return smat object instead of data frame}
}
\value{
if \code{return_smat} - smat object with rows as the intervals.
if not - data frame with the following fields:
'chrom', 'start', 'end', 'cell', 'ncpgs', 'cov', 'meth', 'unmeth', 'avg'
}
\description{
Group CpGs by intervals, and summarise per cell
}
