% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.summarise_by_track}
\alias{smat.summarise_by_track}
\title{Group CpGs by a misha track, and summarise per cell}
\usage{
smat.summarise_by_track(smat, track, breaks, include.lowest = TRUE,
  group_name = "group", parallel = TRUE)
}
\arguments{
\item{smat}{smat object}

\item{track}{misha track expression to summarise by}

\item{breaks}{breaks that determine the bins of the track expression}

\item{include.lowest}{if 'TRUE', the lowest value of the range determined by
breaks is included}

\item{group_name}{name of the column with the group in the returned data frame}

\item{parallel}{compute parallely per group (using doMC package)}
}
\value{
data frame with the following fields:
'group_name', 'cell', 'ncpgs', 'meth', 'unmeth', 'avg'
}
\description{
Group CpGs by a misha track, and summarise per cell
}
