% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.to_df}
\alias{smat.to_df}
\title{export smat object to data frame}
\usage{
smat.to_df(smat, coords = FALSE)
}
\arguments{
\item{smat}{smat object}

\item{coords}{if TRUE - chrom,start,end fields are united to a single field named 'coords'}
}
\value{
intervals (chrom, start, end fields) data frame with additonal fields:
meth (methylated calls), unmeth (unmethylated calls) and cov (total coverage)
}
\description{
export smat object to data frame
}
