% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.to_marginal_track}
\alias{smat.to_marginal_track}
\title{Create track from smat marginals}
\usage{
smat.to_marginal_track(
  smat,
  track,
  description,
  cols = NULL,
  overwrite = FALSE
)
}
\description{
Create track from smat marginals
}
