% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{summarise_intervals}
\alias{summarise_intervals}
\title{Summarise for intervals set}
\usage{
summarise_intervals(db, intervals)
}
\arguments{
\item{db}{cgdb object}

\item{intervals}{intervals set to summarise the genome with. Numeric values would calculate
the summary for equally sized bins.}
}
\value{
coverage and methylated calls for the intervals set.
}
\description{
Summarise for intervals set
}
