% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{cgdb_add_plate}
\alias{cgdb_add_plate}
\title{Add sc5mc data of a plate from smat object}
\usage{
cgdb_add_plate(db, smat, plate_name = NULL, overwrite = TRUE,
  update_cells = TRUE)
}
\arguments{
\item{db}{cgdb object}

\item{smat}{smat object}

\item{plate_name}{name of the plate}

\item{overwrite}{overwrite}

\item{update_cells}{update cells.csv}
}
\description{
Add sc5mc data of a plate from smat object
}
