% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb.R
\name{filter_intervals}
\alias{filter_intervals}
\title{Filter CpGs by genomic intervals}
\usage{
filter_intervals(db, intervals, ...)
}
\arguments{
\item{...}{Arguments passed on to \code{gpatterns::gintervals.filter}
\describe{
  \item{intervals1}{intervals set}
  \item{intervals2}{intervals set by which to filter intervals1}
  \item{max_distance}{maximal distance of every interval in intervals1 from intervals2 (defualt 0)}
  \item{abs_dist}{take the absolute distance}
  \item{bind_intervals2}{cbind add intervals2 to result}
}}
}
\description{
Filter CpGs by genomic intervals
}
