% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-meth.R
\name{get_cgc_trend}
\alias{get_cgc_trend}
\title{Calculate average methylation per cell given CpG content}
\usage{
get_cgc_trend(db, min_cov = 300, breaks = c(seq(0, 0.08, 0.01), 0.15, 1))
}
\arguments{
\item{db}{cgdb object}

\item{min_cov}{minimal number of CpGs for average methylation calculation}

\item{breaks}{breaks of CpG content#'}
}
\value{
cell metadata of \code{db} with additional columns containing the average methylation in 
different CpG content regimes
}
\description{
Calculate average methylation per cell given CpG content
}
