% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgdb-qc.R
\name{qc_plot}
\alias{qc_plot}
\title{Plot QC statistics of cgdb object}
\usage{
qc_plot(db, ofn = NULL, base_height = 12, min_cgc_cov = 300,
  return_all_figs = FALSE, downsample = TRUE, ...)
}
\arguments{
\item{db}{cgdb object}

\item{ofn}{output file}

\item{base_height}{The height (in inches) of the plot or of one sub-plot if \code{nrow}
or \code{ncol} > 1. Default is 4.}

\item{min_cgc_cov}{minimal coverage for average methylation per CpG content estimation}

\item{return_all_figs}{return a list of all the figures}

\item{downsample}{downsample cells when plotting global trends}

\item{...}{additional parameters for cowplot::save_plot}
}
\description{
Plot QC statistics of cgdb object
}
