% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{sc5mc.generate_metacells}
\alias{sc5mc.generate_metacells}
\title{Genarate metacells from TAD methylation}
\usage{
sc5mc.generate_metacells(
  db,
  tad_intervs,
  knn = 80,
  k_expand = 10,
  min_cluster_size = 20,
  n_resamp = 1000,
  seed = NULL,
  K = 50,
  min_mc_size = 30,
  min_tad_cov = 10,
  min_tad_cells = 50,
  min_tads = 1000,
  tads_meth = NULL
)
}
\arguments{
\item{db}{cgdb object}

\item{tad_intervs}{intervals set of TADs}

\item{knn}{knn}

\item{k_expand}{k_expand}

\item{min_cluster_size}{min_cluster_size}

\item{n_resamp}{n_resamp}

\item{seed}{seed}

\item{K}{K}

\item{min_mc_size}{min_mc_size}

\item{min_tad_cov}{min_tad_cov}

\item{min_tad_cells}{min_tad_cells}

\item{min_tads}{min_tads}

\item{tads_meth}{previously computed methylation of TADS (using get_tads_meth)}
}
\value{
list with the following fields: mc_map, tads_clust, tads_meth, tads_meth_f, tads_mc
}
\description{
Genarate metacells from TAD methylation
}
