% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{smat.calc_pdiff_cor}
\alias{smat.calc_pdiff_cor}
\title{calculate pairwise correlation of cells pdiff}
\usage{
smat.calc_pdiff_cor(smat, min_cells = 10, min_cgs = 100,
  use = "pairwise.complete.obs", method = "spearman", intervs = NULL,
  cols = NULL, samp_data = NULL)
}
\arguments{
\item{smat}{smat object}

\item{min_cells}{minimal number of cells with pdiff statistic in both cells}

\item{min_cgs}{minimal number of cpgs covered in both cells}

\item{use}{`use` parameter of `cor` function}

\item{method}{`method` parameter of `cor` function}

\item{intervs}{intervals set to filter smat by before calculating. if NULL
all CpGs would be used.}

\item{cols}{vector of names of cells to use. If NULL all cells would be used.}

\item{samp_data}{data frame with additional annotation of the cell.
must have 'track' field with the cell name.}
}
\value{
smat object with cell_cor field
}
\description{
calculate pairwise correlation of cells pdiff
}
