% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.multi_join}
\alias{smat.multi_join}
\title{join multiple smat objects}
\usage{
smat.multi_join(smat_list, type = "full")
}
\arguments{
\item{smat_list}{list of smat objects}

\item{type}{join type. can be 'inner', 'full', 'left', 'right'. see xx_join by dplyr package.}
}
\value{
joined smat object
}
\description{
join multiple smat objects
}
