% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{smat.plot_cor_mat}
\alias{smat.plot_cor_mat}
\title{plot cell cell correlation matrix}
\usage{
smat.plot_cor_mat(
  smat,
  row_ord = NULL,
  col_ord = NULL,
  show_colnames = FALSE,
  show_rownames = FALSE,
  breaks = NULL,
  color_pal = NULL,
  title = TRUE,
  ...
)
}
\arguments{
\item{smat}{smat object}

\item{row_ord}{specific order for the rows}

\item{col_ord}{specific order for the columns}

\item{show_colnames}{show column names}

\item{show_rownames}{show row names}

\item{color_pal}{color pallete}

\item{title}{show title}

\item{...}{other parameters of pheatmap1}
}
\description{
plot cell cell correlation matrix
}
