% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smat.R
\name{smat.summarise_cpgs}
\alias{smat.summarise_cpgs}
\title{Summarises CpGs by groups pf cells}
\usage{
smat.summarise_cpgs(smat, groups_df = NULL, groups = NULL, min_cells = 1,
  max_cells = Inf)
}
\arguments{
\item{smat}{smat object}
}
\value{
tidy data frame with summariy statistics for each group and CpG
}
\description{
Summarises CpGs by groups pf cells
}
